+++ src/edit/vi.rs
@@ -272,11 +272,11 @@ impl<'a> Edit for ViEditor<'a> {
         self.editor.set_select_opt(select_opt);
     }
 
-    fn tab_width(&self) -> usize {
+    fn tab_width(&self) -> u16 {
         self.editor.tab_width()
     }
 
-    fn set_tab_width(&mut self, tab_width: usize) {
+    fn set_tab_width(&mut self, tab_width: u16) {
         self.editor.set_tab_width(tab_width);
     }
